Add mcollective::stomp_failover_pool class
This class provides an interface for describing something that can be
hard to do using the Puppet DSL - related collections of values with
indices. It was the one edge case mcollective configuration problem I
was aware of that the basic type/provider mcollective_setting couldn't
easily handle.

Update README to show usage of new type